+++ core/jobs/src/test/java/com/cognifide/aet/job/common/modifiers/resolution/ResolutionModifierTest.java
@@ -22,6 +22,8 @@ import static org.mockito.Mockito.when;
 
 import com.cognifide.aet.job.api.exceptions.ParametersException;
 import com.cognifide.aet.job.api.exceptions.ProcessingException;
+import com.cognifide.aet.job.common.utils.javascript.JavaScriptJobExecutor;
+import com.cognifide.aet.job.common.utils.javascript.JavaScriptJobResult;
 import java.util.Map;
 import org.junit.Before;
 import org.junit.Test;
@@ -72,6 +74,9 @@ public class ResolutionModifierTest {
 
   @Mock
   private Dimension windowDimension;
+
+  @Mock
+  private JavaScriptJobExecutor jsExecutor;
 
   @InjectMocks
   private ResolutionModifier tested;
@@ -149,8 +154,8 @@ public class ResolutionModifierTest {
   @Test
   public void collectTest_setOnlyWidth() throws ParametersException, ProcessingException {
     setUp_setOnlyWidth();
-    when(webDriver.executeScript(JAVASCRIPT_GET_BODY_HEIGHT))
-        .thenReturn(CUSTOM_HEIGHT);
+    when(jsExecutor.execute(JAVASCRIPT_GET_BODY_HEIGHT))
+        .thenReturn(new JavaScriptJobResult(CUSTOM_HEIGHT));
 
     tested.setParameters(params);
     tested.collect();
@@ -161,10 +166,11 @@ public class ResolutionModifierTest {
   }
 
   @Test
-  public void collectTest_setOnlyWidth_reachedBrowserHeightLimit() throws ParametersException, ProcessingException {
+  public void collectTest_setOnlyWidth_reachedBrowserHeightLimit()
+      throws ParametersException, ProcessingException {
     setUp_setOnlyWidth();
-    when(webDriver.executeScript(JAVASCRIPT_GET_BODY_HEIGHT))
-        .thenReturn(BROWSER_HEIGHT_LIMIT + 5000L);
+    when(jsExecutor.execute(JAVASCRIPT_GET_BODY_HEIGHT))
+        .thenReturn(new JavaScriptJobResult(BROWSER_HEIGHT_LIMIT + 5000L));
 
     tested.setParameters(params);
     tested.collect();