+++ integration-tests/sanity-functional/src/test/java/com/cognifide/aet/sanity/functional/HomePageTilesTest.java
@@ -29,15 +29,15 @@ import org.junit.runner.RunWith;
 @Modules(GuiceModule.class)
 public class HomePageTilesTest {
 
-  private static final int TESTS = 140;
+  private static final int TESTS = 147;
 
-  private static final int EXPECTED_TESTS_SUCCESS = 79;
+  private static final int EXPECTED_TESTS_SUCCESS = 84;
 
   private static final int EXPECTED_TESTS_CONDITIONALLY_PASSED = 11;
 
   private static final int EXPECTED_TESTS_WARN = 5;
 
-  private static final int EXPECTED_TESTS_FAIL = 56;
+  private static final int EXPECTED_TESTS_FAIL = 58;
 
   @Inject
   private ReportHomePage page;